refactor: move interface into sectionTypes/
