+++ src/schema/v2/homeView/sectionTypes/HomeViewExploreBySectionType.ts
@@ -6,7 +6,7 @@ import {
 } from "graphql"
 import { ResolverContext } from "types/graphql"
 import { NodeInterface } from "../../object_identification"
-import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
+import { HomeViewGenericSectionInterface } from "./HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "./standardSectionFields"
 import { ImageType } from "../../image"
