+++ src/schema/v2/homeView/sectionTypes/HomeViewFairsSectionType.ts
@@ -3,7 +3,7 @@ import { pageable } from "relay-cursor-paging"
 import { ResolverContext } from "types/graphql"
 import { fairsConnection } from "../../fairs"
 import { NodeInterface } from "../../object_identification"
-import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
+import { HomeViewGenericSectionInterface } from "./HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "./standardSectionFields"
