+++ src/schema/v2/homeView/sectionTypes/HomeViewHeroUnitsSectionType.ts
@@ -4,7 +4,7 @@ import { ResolverContext } from "types/graphql"
 import { emptyConnection } from "../../fields/pagination"
 import { heroUnitsConnection } from "../../HeroUnit/heroUnitsConnection"
 import { NodeInterface } from "../../object_identification"
-import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
+import { HomeViewGenericSectionInterface } from "./HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
 import { standardSectionFields } from "./standardSectionFields"
